requirements.txt
Flask>=2.0,<3.0
requests>=2.28
PyYAML>=6.0
python-dotenv>=0.21
